Move the general logger code to a dedicated module.
